Refactored createServer and database.ts to allow for composition and testing that works.
Couldn't do pre-push hooks with testing because tests would hang due to unclosed database
handles. This sparked a whole series of refactoring of createServer to now take in a parameter
of sessionStore (because need to be able to close the store externally), getting rid of
setupTests, and adding in constructor logic for MongoConnection so that it can more cleanly
handle having an in memory vs actual database connection.